ai-job-agent/requirements.txt
python-jobspy==1.1.80
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
